feat: Move ParticipantsList to a separate File
